+++ tests/xlsx/test_excel_writer.py
@@ -95,7 +95,7 @@ def test_write_xlsx__runs_beta_1_file_without_error():
 
 
 @pytest.mark.slow
-def test_write_xlsx__runs_magnet_finding_alg_without_error():
+def test_write_xlsx__runs_magnet_finding_alg_without_error(patch_get_positions):
     # Tanner (12/8/22): do not add anything to this test, it is just meant to run a full analysis start to
     # finish with no mocking. This is specifically to make sure that there are no issues with using the magnet
     # finding alg since it is often mocked in other tests to make them run faster.
@@ -103,7 +103,7 @@ def test_write_xlsx__runs_magnet_finding_alg_without_error():
     # done in the tests below.
 
     pr = PlateRecording(TEST_FILE_PATH)
-    output_file_name = write_xlsx(pr)
+    output_file_name = write_xlsx(pr, stim_waveform_format="stacked")
 
     # this assertion isn't really necessary, but it's nice to make an assertion in a test that otherwise has none
     assert isinstance(output_file_name, str)